+++ src/ui/sound/misc.rs
@@ -1,9 +1,10 @@
 //! Section containing misc parameters
 
-use iced::{slider, Column, Container, Element, Text};
+use iced::{Column, Container, Element, Text};
 
 use crate::messages::Message;
 use crate::params::{GetValue, SoundParameter, SoundParameterValues};
+use crate::ui::elements::slider_widget as slider;
 use crate::ui::elements::{checkbox::checkbox_with_labels, slider::slider_with_labels};
 use crate::ui::style;